- Code is segregated into different methods and in different files.
