+++ TWB-Ecommerce/Features/Bag/Models/BagItem.swift
@@ -1,0 +1,42 @@
+//
+//  BagItem.swift
+//  TWB-Ecommerce
+//
+//  Created by Hassam Ali on 19/09/2024.
+//
+
+import Foundation
+
+struct BagViewItemModel: Equatable, Hashable, Identifiable {
+    let id: String
+    let price: Float
+    var count: Int
+    let letters: String?
+    let flavour: String?
+    let topper: String
+    let customisedTopper: Bool
+    let vat: Float // Question: is it for a single product or for all product. Like count is 2, then for 2 or for 1
+    let imageURL: String
+    var totalPrice: Float {
+        (Float(count) * price) + vat
+    }
+}
+
+let items: [BagViewItemModel] = [.init(id: "Tube Acrylic 018",
+                                       price: 447,
+                                       count: 1,
+                                       letters: "A",
+                                       flavour: "Vanilla Mistikkah",
+                                       topper: "Congratulations (AED 32",
+                                       customisedTopper: false,
+                                       vat: 10,
+                                       imageURL: ""),
+                                 .init(id: "Long Acrylic Tray 006",
+                                       price: 1053,
+                                       count: 2,
+                                       letters: "A",
+                                       flavour: "Vanilla Mistikkah",
+                                       topper: "Congratulations (AED 32",
+                                       customisedTopper: false,
+                                       vat: 10,
+                                       imageURL: "")]
